Improve organization of GPP Criteria domain knowledge (#26)
This PR solves the issues raised in #25 but not in the approach that was initially suggested.

To make the GPP criteria domain knowledge more intuitive, only a couple of extra arguments remained: Description and Selection Criteria Type (only applicable to SC).

The rest of the fields are computed from the fields already available in the GPP Criteria.